+++ Projects/Features/HomeFeature/Sources/Presentation/Home/UI/Cell/ListKitBestPopupCell.swift
@@ -1,0 +1,60 @@
+//
+//  ListKitBestPopupCell.swift
+//  HomeFeature
+//
+//  Created by 김동현 on 6/15/26.
+//
+
+import SwiftUI
+import Domain
+import Kingfisher
+import DSKit
+
+struct ListKitBestPopupCell: View {
+    let popup: Popup
+
+    var body: some View {
+        ZStack(alignment: .bottomLeading) {
+            KFImage(URL(string: popup.imageUrlList.first ?? ""))
+                .downSampled(.bestPopupCell)
+                .scaledToFill()
+                .frame(width: 194, height: 271)
+                .clipped()
+
+            LinearGradient(
+                gradient: Gradient(stops: [
+                    .init(color: Color.mainBlack.opacity(0.0), location: 0.00),
+                    .init(color: Color.mainBlack.opacity(0.50), location: 0.52),
+                    .init(color: Color.mainBlack.opacity(1.00), location: 0.83),
+                ]),
+                startPoint: .top,
+                endPoint: .bottom
+            )
+            .frame(height: 100)
+            .clipped()
+
+            VStack(alignment: .leading) {
+                Text(popup.name)
+                    .font(.scdream(.bold, size: 15))
+                    .foregroundStyle(Color.bestPostTitle)
+                    .lineLimit(1)
+                    .truncationMode(.tail)
+
+                HStack(spacing: 2) {
+                    DSKitResource.image("Address")
+                        .resizable()
+                        .renderingMode(.template)
+                        .aspectRatio(contentMode: .fit)
+                        .foregroundStyle(Color.bestPostAddress)
+                        .frame(width: 15, height: 15)
+
+                    Text(popup.roadAddress.shortAddress)
+                        .font(.scdream(.medium, size: 12))
+                        .foregroundStyle(Color.bestPostAddress)
+                }
+            }
+            .padding(11)
+        }
+        .contentShape(Rectangle())
+    }
+}
