+++ Projects/Features/HomeFeature/Sources/Presentation/Home/UI/Cell/ListKitGridPopupCell.swift
@@ -1,0 +1,68 @@
+//
+//  ListKitGridPopupCell.swift
+//  HomeFeature
+//
+//  Created by 김동현 on 6/15/26.
+//
+
+import SwiftUI
+import Domain
+import Kingfisher
+import DSKit
+
+struct ListKitGridPopupCell: View {
+    let popup: Popup
+    let isLiked: Bool
+    let cellWidth: CGFloat
+    let toggleLike: () -> Void
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 0) {
+            ZStack(alignment: .topTrailing) {
+                KFImage(URL(string: popup.imageUrlList.first ?? ""))
+                    .placeholder {
+                        Rectangle()
+                            .fill(Color.subWhite)
+                            .frame(width: cellWidth, height: 217)
+                    }
+                    .downSampled(.gridPopupCell)
+                    .scaledToFill()
+                    .frame(width: cellWidth, height: 217)
+                    .clipped()
+
+                Button {
+                    toggleLike()
+                } label: {
+                    DSKitResource.image(isLiked ? "favorite_fill" : "favorite")
+                        .renderingMode(.template)
+                        .resizable()
+                        .aspectRatio(contentMode: .fit)
+                        .frame(width: 25, height: 25)
+                        .foregroundStyle(isLiked ? Color.mainOrange : Color.subWhite)
+                }
+                .padding(10)
+                .applyShadow(color: .mainBlack, alpha: 0.25, x: 0, y: 1, blur: 3)
+            }
+            .frame(width: cellWidth, height: 217)
+
+            Text(popup.roadAddress.shortAddress)
+                .font(.scdream(.regular, size: 12))
+                .foregroundStyle(Color.mainBlack)
+                .padding(.top, 10)
+
+            Text(popup.name)
+                .font(.scdream(.medium, size: 14))
+                .foregroundStyle(Color.mainBlack)
+                .lineLimit(1)
+                .truncationMode(.tail)
+                .padding(.top, 5)
+
+            Text("\(popup.startDate, formatter: DateFormatter.popupDateFormat) - \(popup.endDate, formatter: DateFormatter.popupDateFormat)")
+                .ppStyleFontFixedSpacing(.scdream(.regular, size: 12), letterSpacingPt: -1)
+                .foregroundStyle(Color.mainGray)
+                .padding(.top, 5)
+        }
+        .frame(width: cellWidth, alignment: .leading)
+        .contentShape(Rectangle())
+    }
+}
